false start

+++ src/net/mlin/vcfGLuer/joint/JointCalling.kt
@@ -77,7 +77,7 @@ fun jointCall(
         },
         RowEncoder.apply(
             StructType()
-                .add("variantId", DataTypes.IntegerType, false)
+                .add("variantId", DataTypes.LongType, false)
                 .add("sampleId", DataTypes.IntegerType, false)
                 .add("entry", DataTypes.StringType, false)
         )
@@ -85,9 +85,9 @@ fun jointCall(
 
     // group the genotype entries by variantId and generate pVCF text lines (Snappy-compressed)
     val pvcfLinesDF = entriesDF
-        .groupByKey(MapFunction<Row, Int> { it.getAs<Int>(0) }, Encoders.INT())
+        .groupByKey(MapFunction<Row, Long> { it.getAs<Long>(0) }, Encoders.LONG())
         .mapGroups(
-            MapGroupsFunction<Int, Row, Row> { variantId, variantEntries ->
+            MapGroupsFunction<Long, Row, Row> { variantId, variantEntries ->
                 // look up this variant in the broadcast database file (using pooled resources...)
                 ensureVariantsDb(variantsDbLocalFilename, variantsDbSharedFilename)
                 val variant = ExitStack().use { cleanup ->
@@ -97,7 +97,7 @@ fun jointCall(
                     val getVariant = cleanup.add(
                         variants.prepareStatement("SELECT * from Variant WHERE variantId = ?")
                     )
-                    getVariant.setInt(1, variantId)
+                    getVariant.setLong(1, variantId)
                     val rs = cleanup.add(getVariant.executeQuery())
                     check(rs.next())
                     Variant(rs)
@@ -110,11 +110,12 @@ fun jointCall(
                     variant,
                     variantEntries
                 )
+                // TODO: assign region
                 RowFactory.create(variantId, snappyLine)
             },
             RowEncoder.apply(
                 StructType()
-                    .add("variantId", DataTypes.IntegerType, false)
+                    .add("variantId", DataTypes.LongType, false)
                     .add("snappyLine", DataTypes.BinaryType, false)
             )
             // persist before sorting: https://stackoverflow.com/a/56310076
@@ -131,7 +132,7 @@ fun jointCall(
     return pvcfHeader to pvcfLinesDF
         .toJavaRDD()
         .sortBy(
-            Function<Row, Int> { it.getAs<Int>(0) },
+            Function<Row, Long> { it.getAs<Long>(0) },
             true,
             // Coalesce to fewer partitions now that the data size has been reduced, to output a
             // smaller number of reasonably-sized pVCF part files.
@@ -187,7 +188,7 @@ fun generateJointCalls(
     variantsDbFilename: String,
     callsetId: Int,
     vcfFilename: String
-): Sequence<Triple<Int, Int, String>> = // [(variantId, sampleId, pvcfEntry)]
+): Sequence<Triple<Long, Int, String>> = // [(variantId, sampleId, pvcfEntry)]
     sequence {
         // variants sequence (read from broadcast database file)
         val variants = sequence {
@@ -199,7 +200,7 @@ fun generateJointCalls(
                     "SELECT * FROM Variant ORDER BY variantId"
                 )
                 while (rs.next()) {
-                    val variantId = rs.getInt("variantId")
+                    val variantId = rs.getLong("variantId")
                     val variant = Variant(rs)
                     yield(variantId to variant)
                 }
@@ -232,7 +233,7 @@ fun generateJointCalls(
  * Callset records assorted into reference bands, records containing the focal variant, and others
  */
 class GenotypingContext(
-    val variantId: Int,
+    val variantId: Long,
     val variant: Variant,
     val callsetRecords: List<VcfRecordUnpacked>
 ) {
@@ -260,7 +261,7 @@ class GenotypingContext(
  * too overlapping. These assumptions match up with small-variant gVCF inputs, of course.
  */
 fun generateGenotypingContexts(
-    variants: Sequence<Pair<Int, Variant>>, // (variantId, variant)
+    variants: Sequence<Pair<Long, Variant>>, // (variantId, variant)
     recordsIter: Iterator<VcfRecord>
 ): Sequence<GenotypingContext> {
     // buffer of records whose ranges don't strictly precede (<<) the last-processed variant, nor
